Add __exit__ flag handling to r20; beautify.
Relax the requirement that it be running in a virtual
environment as well. It is paramount that r20 *works*
so that the "python -m r20 __exit__ <val>" call will
prove to the caller that r20 is online for them.

Therefor we want the script to be simple and failsafe.
We can source out testing for virtual environments to
other scripts.